public-code/requirements.txt
numpy
pandas==1.2.1
tqdm
Pillow
scipy==1.1.0
imageio